requirements.txt
python-socketio==5.0.4
structlog==25.4.0